Now check Network post
Now, the add on check if the Blacklisted words are in the sub site post
title, content, permanlink and the Author name, blog name (the splogger
can change the name after signup).

If a blacklisted word is detected, an email is sent to Network Admin
with the sub site url and post url.